+++ VariationalBayes/__init__.py
@@ -1,2 +1,1 @@
 from VariationalBayes import Parameters
-    # VectorParam, ScalarParam, PosDefMatrixParam, ModelParamsDict
